+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
+## Unreleased
+### Fixed
+* changed default value of `hasStaticData` to `false`
+
 ## [2.10.2] - 04-10-2018
 ### Fixed 
 * set Content-Type: application/javascript when response is wrapped in callback (JSONP)